refactor: model and view seperate
