+++ src/main/java/racingcar/view/InputViewCarName.java
@@ -1,0 +1,18 @@
+package racingcar.view;
+
+import java.util.Scanner;
+
+public class InputViewCarName extends InputView<String[]> {
+
+	public InputViewCarName(String text) {
+		super(text);
+	}
+
+	@Override
+	public String[] getInput() {
+		super.printText();
+		Scanner sc = new Scanner(System.in);
+		String input = sc.nextLine();
+		return input.split(",");
+	}
+}
